+++ resource/logresource/error.go
