+++ resource/logresource/resource.go
