+++ resource/logresource/resource_test.go
