+++ resource/logresource/wrap.go
