+++ resource/metricsresource/resource_test.go
